App/requirements.txt
fastapi == 0.104.1
uvicorn == 0.23.2
pydantic == 1.10.8
requests ==   2.31.0
google-auth == 2.23.4
google-auth-oauthlib == 1.1.0 
starlette == 0.27.0
itsdangerous == 2.1.2
bcrypt == 4.0.1
PyJWT == 2.8.0
python-multipart == 0.0.6
pandas === 2.1.3
scikit-learn == 1.3.2
joblib == 1.3.2
xgboost ==  2.0.2